bugfix: uint16_t is too small for posarray->size and posarray->next
Change the type of posarray->size and posarray->next from uint16_t
to uint32_t. 16 bits are not enough to hold really large messages.

+++ entry.h
@@ -15,8 +15,8 @@
 #include "khash.h"
 
 typedef struct posarray {
-  uint16_t size;
-  uint16_t next;
+  uint32_t size;
+  uint32_t next;
   pos_t* data;
 } posarray;
